Code refactoring. Complete and test user management

+++ models/form/UserForm.php
@@ -1,14 +1,17 @@
 <?php
 
-namespace models;
+namespace models\form;
 
 use core\FormModel;
 use core\Application;
+use core\Controller;
 use core\FormModelField;
 use core\SqlModel;
-use models\UserModel;
 
-class UserWorkerFormModel extends FormModel
+use models\sql\User;
+use models\sql\Worker;
+
+class UserForm extends FormModel
 {
     const DB_TABLE = 'user';
     const FORM_SUBMIT_VALUE = 'Update User';
@@ -47,21 +50,21 @@ class UserWorkerFormModel extends FormModel
             model: $this,
             type: FormModelField::TYPE_TEXT,
             label: 'First Name',
-            rules: [self::RULE_REQUIRED]
+            rules: [self::RULE_REQUIRED, [self::RULE_MAX, 'max' => 100]]
         );
         $this->last_name = new FormModelField(
             name: "last_name",
             model: $this,
             type: FormModelField::TYPE_TEXT,
             label: 'Last Name',
-            rules: [self::RULE_REQUIRED]
+            rules: [self::RULE_REQUIRED, [self::RULE_MAX, 'max' => 100]]
         );
         $this->phone_number = new FormModelField(
             name: "phone_number",
             model: $this,
             type: FormModelField::TYPE_TEXT,
             label: 'Phone Number',
-            rules: [self::RULE_REQUIRED]
+            rules: [self::RULE_REQUIRED, [self::RULE_MAX, 'max' => 20]]
         );
         $this->admin = new FormModelField(
             name: "admin",
@@ -106,7 +109,7 @@ class UserWorkerFormModel extends FormModel
             rules: []
         );
 
-        if(!Application::isAdmin()) {
+        if (!Controller::isUserAdmin()) {
             $this->admin->rules[] = self::RULE_READONLY;
             $this->worker->rules[] = self::RULE_READONLY;
             $this->supervisor->rules[] = self::RULE_READONLY;
@@ -119,7 +122,7 @@ class UserWorkerFormModel extends FormModel
     {
         parent::loadDataFromSqlModel($model);
 
-        if ($model instanceof UserModel) {
+        if ($model instanceof User) {
             $this->worker->value = $model->worker;
             if ($model->worker) {
                 $this->supervisor->value = $model->worker->supervisor;
@@ -131,37 +134,43 @@ class UserWorkerFormModel extends FormModel
         }
     }
 
-    public function sendDataToSqlModel(SqlModel $model, array $toIgnore = ['worker'])
+    public function loadDataFromArray($arr)
+    {
+        parent::loadDataFromArray($arr);
+
+        if (!$this->worker->value) {
+            $this->fieldsToIgnore = ['supervisor', 'can_drive', 'special_effects_license'];
+        }
+    }
+
+    public function loadDataToSqlModel(SqlModel $model, array $toIgnore = ['worker'])
     {
 
-        if ($model instanceof UserModel) {
+        if ($model instanceof User) { 
             if ($this->worker->value and !$model->worker) { // if we set the user as a worker but they are not a worker yet
-                $model->worker = new WorkerModel();
+                $model->worker = new Worker();
                 // We try to see if the user was previously a worker
-                $wasWorker = WorkerModel::getByWhere('user_id = ?', [$model->id]);
-                if(count($wasWorker) !== 0) { // If the user was previously a worker
+                $wasWorker = Worker::getByWhere('user_id = ?', [$model->id]);
+                if (count($wasWorker) !== 0) { // If the user was previously a worker
                     $model->worker = $wasWorker[0];
                     $model->worker->status = 1;
-                }
-                else { // If the user was not previously a worker
+                } else { // If the user was not previously a worker
                     $model->worker->status = 1;
                     $model->worker->user_id = $model->id;
                     $model->worker->supervisor = $this->supervisor->value;
                     $model->worker->can_drive = $this->can_drive->value;
                     $model->worker->special_effects_license = $this->special_effects_license->value;
                 }
-            }
-            else if (!$this->worker->value and $model->worker) { // If we unset the user as a worker but they are a worker
+            } else if (Controller::isUserAdminOrSupervisor() and !$this->worker->value and $model->worker) { // If we unset the user as a worker but they are a worker
                 $model->worker->delete();
                 $model->worker = false;
-            }
-            else if ($this->worker->value and $model->worker) { // If we update the worker attributes
+            } else if (Controller::isUserAdminOrSupervisor() and $this->worker->value and $model->worker) { // If we update the worker attributes
                 $model->worker->supervisor = $this->supervisor->value;
                 $model->worker->can_drive = $this->can_drive->value;
                 $model->worker->special_effects_license = $this->special_effects_license->value;
             }
         }
 
-        parent::sendDataToSqlModel($model, $toIgnore);
+        parent::loadDataToSqlModel($model, $toIgnore);
     }
 }
